requirements.txt
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0
web3==6.11.3
